Proper end date and start date support
If a start date or end date is not specified, we'll use the "memberSince" date
and the last sync date from Fitbit for star and end dates.

Also don't grab data for days after the last sync date - 1 day. This helps
account for days that don't have data or have only partial data. Ex. I sync
at 11am on 2013-08-20 and don't want to record 2013-08-20 until I have the
complete days worth of data and sync on 2013-08-21 or later.